backend/requirements.txt
Django==4.1.13
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
celery==5.3.6
redis==4.5.5
django-celery-beat==2.5.0
django-celery-results==2.5.1
mysqlclient==2.2.4
python-dotenv==1.0.0
drf-yasg==1.21.7
requests==2.31.0
importlib-metadata==6.8.0
click>=8.1.2
setuptools>=69.0.3
packaging>=23.2
Pillow==10.1.0